UniADet-repo/requirements.txt
numpy
Pillow
scikit-learn
torch
torchvision
tqdm
transformers>=4.37.0

